Make the audio module less unsafe
This is achieved by eliminating two global variables and making the code
work nicely with ownership analysis.

This is possible because AudioDevice<NesCallback> has a lock method that
returns an object that dereferences to NesCallback we can store the
output buffer directly in NesCallback and then safely access it to
update the raw sound data.

The resulting code is easier to follow and contains fewer unsafe blocks.